wvcc compiler macros now generate explicit CC and CXX scripts.
This gets rid of the need for "make VERBOSE=1", at least for the
compilation stage if not the linking stage.  The commands look like a
simple "CC -c filename", and if you run that from the prompt, it'll
do exactly the same thing as when make does it.

Borrowed this idea (but not implementation) from djb's Makefiles.